Moved ranking table to separate component
